Refactor Project Structure
